Added FileSystem, MemoryFS, and test setup.
Created the FileSystem interface and moved the spec there from
Clerk.go. Wrote unit tests for any implementation of FileSystem.
Created, but did not implement, MemoryFS, and wrote unit tests
for it that call the FileSystem tests. Changed Clerk to implement
FileSystem and wrote FileSystem tests for it on various difficulties
in clerk_test.go.